chore: Move type into consumer package.
