+++ internal/codeintel/autoindexing/dependency_sync_scheduler.go
@@ -10,7 +10,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/codeintel/autoindexing/shared"
 	"github.com/sourcegraph/sourcegraph/internal/codeintel/dependencies"
-	codeinteltypes "github.com/sourcegraph/sourcegraph/internal/codeintel/types"
 	uploadsshared "github.com/sourcegraph/sourcegraph/internal/codeintel/uploads/shared"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/conf/reposource"
@@ -69,7 +68,7 @@ func (h *dependencySyncSchedulerHandler) Handle(ctx context.Context, logger log.
 		return nil
 	}
 
-	job := record.(codeinteltypes.DependencySyncingJob)
+	job := record.(shared.DependencySyncingJob)
 
 	scanner, err := h.uploadsSvc.ReferencesForUpload(ctx, job.UploadID)
 	if err != nil {
