+++ internal/codeintel/autoindexing/shared/workerutil.go
@@ -1,4 +1,4 @@
-package types
+package shared
 
 import "time"
